requirements.txt
django
channels[daphne]
channels_redis
djangorestframework
psycopg2